% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectancy_hp.R
\name{expectancy}
\alias{expectancy}
\alias{expectancy.HP}
\title{HP Model: Life expectancy}
\usage{
expectancy(fit, Ex = NULL, age = NULL, graph = TRUE,
 max_age = 110, prob = 0.95)
}
\arguments{
\item{fit}{Object of the class 'HP' or 'ClosedHP' fitted by the hp() or hp_close() functions.}

\item{Ex}{Numeric vector with the exposure by age. This argument is only necessary when using poisson and binomial models. The default value is NULL.}

\item{age}{Numeric vector specifying the ages to calculate the life expectancy. The default is a sequence (0, 10, 20, ...) until the last decade used in the fitted model.}

\item{graph}{Logical value (TRUE ou FALSE). If TRUE, it also returns a plot. The default value is TRUE.}

\item{max_age}{Positive number indicating  the last age to be used in the calculation of the life expectancy. The default value is 110.}

\item{prob}{A number specifying the probability of credible interval The default value is 0.95.}
}
\value{
A data.frame and (if graph = TRUE) a plot.
}
\description{
This function calculates the life expectancy for each age in fitted HP model.
}
\details{
This function calculates the life expectancy using the truncated life expectancy formula:

 \eqn{e_x = \sum tp_x}

where:

 \eqn{tp_x = p_0 x p_1 x ... x p_x}
}
\examples{
## Importing mortality data from USA
data(USA)

# Example 1: --------------------------------

USA1990 = USA[USA$Year == 1990,]

Ex = USA1990$Ex.Total[1:91]
Dx = USA1990$Dx.Total[1:91]
x = 0:90

fit <- hp(x, Ex, Dx, model = "binomial")
exp <- expectancy(fit)

exp$plot
exp$tabua
exp


# Example 2: -------------------------------

# Using some arguments:

Ex = USA1990$Ex.Total[1:106]

exp <- expectancy(fit, Ex = Ex, age = c(0,20,30,60,105),
                  max_age = 105, prob = 0.99, graph = FALSE)

exp

}
