% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_hp.R
\name{Heatmap}
\alias{Heatmap}
\alias{Heatmap.HP}
\title{HP Model: Heatmap for the life expectancy.}
\usage{
Heatmap(fits, x_lab, age = 0:90, max_age = 110, color = c("red","white","blue"))
}
\arguments{
\item{fits}{Object or a list of objects of the class 'HP' or 'ClosedHP' returned by the hp() or close_hp() functions.}

\item{x_lab}{Vector with the identification of each object in "fits", in the same order.}

\item{age}{Vector with the ages to plot the heatmap.}

\item{max_age}{Positive number indicating  the last age to be used in the calculation of the life expectancy. The default value is 110.}

\item{color}{A length 3 vector with the colors used in the heatmap.}
}
\value{
Heatmap of the life expectancy.
}
\description{
This function plots a heatmap for the life expectancy of the fitted HP models.
}
\examples{
## Importing mortality data from USA
data(USA)

# Example 1: -------------------------------

USA2010 = USA[USA$Year == 2010,]

ExF = USA2010$Ex.Female[1:91]
DxF = USA2010$Dx.Female[1:91]
x <- 0:90

fitF <- hp(x, ExF, DxF, model = "lognormal")

Heatmap(fitF, x_lab = "Female expec. 2010 USA")


# Example 2: -------------------------------

# A plot with more than one fitted model:

ExM = USA2010$Ex.Male[1:91]
DxM = USA2010$Dx.Male[1:91]

fitM <- hp(x, ExM, DxM, model = "lognormal")

fits <- list(fitF = fitF, fitM = fitM)

Heatmap(fits, x_lab = c("Female 2010 USA","Male 2010 USA"),
        age = 15:85, max_age = 100, color = c("pink","black","green"))

}
