% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp.R
\name{hp}
\alias{hp}
\title{Bayesian HP}
\usage{
hp(x, Ex, Dx, model = c("binomial", "lognormal", "poisson"),
 M = 50000, bn = round(M/5), thin = 10, m = rep(NA, 8), v = rep(NA, 8),
 inits = NULL, K = NULL, sigma2 = NULL, prop.control = NULL,
 reduced_model = FALSE)
}
\arguments{
\item{x}{Numeric vector of the ages.}

\item{Ex}{Numeric vector with the exposure by age.}

\item{Dx}{Numeric vector with the death counts by age.}

\item{model}{Character string specifying the model to be used. The options are: "binomial", "lognormal" or "poisson".}

\item{M}{Positive integer indicating the number of iterations of the MCMC run.}

\item{bn}{Nonnegative integer indicating the number of iteration to be discarded as the burn-in period.}

\item{thin}{Positive integer specifying the period for saving samples.}

\item{m}{Numeric vector with the mean of the prior distributions for (A, B, C, D, E, F, G, H).}

\item{v}{Numeric vector with the variance of the prior distributions for (A, B, C, D, E, F, G, H).}

\item{inits}{Numeric vector with the initial values for the parameters (A, B, C, D, E, F, G, H).}

\item{K}{Number that specifies the extra paramater 'K' value for binomial and poisson models. It is used only if model = "binomial" or model = "poisson".}

\item{sigma2}{Positive number that specifies initial value of sigma2 in lognormal model. It is used only if model = "lognormal".}

\item{prop.control}{Positive number that is used for tuning the acceptance rate of MCMC.}

\item{reduced_model}{Logical value which determines if reduced model should be used.}
}
\value{
A HP class object.
\item{summary}{A table with summaries measures of the parameters.}
\item{post.samples}{A list with the chains generated by MCMC.}
\item{data}{A table with the data used in fitted model.}
\item{info}{A list with some informations of the fitted model like prior distributions mean and variance, initial values.}
}
\description{
This function fits a HP model following a Bayesian framework using MCMC methods to sample
from the parameters' posterior distribution. There are three model options. The Binomial
and Poisson models use the 9 parameters HP curve. The lognormal model uses the traditional
8 parameters HP curve. The extra parameter 'K' from 9 parameters curve value is estimated
using an optimization function. The other parameters are estimated by the MCMC method using
the following prior distributions: Beta for A, B, C, D and G parameters. Gamma for E and H.
Normal for F parameter and Inverse Gamma for sigma2.
}
\details{
The binomial model assumes that Dx, the death count for the age x, has a binomial distribution:
Bin(Ex, qx), where qx is probability of death in age x. The poisson model assumes that Dx has a
Poisson distribution: Po(Ex*qx). These models use the 9 parameters
HP curve, that was proposed by Heligman and Pollard (1980):

 \eqn{HP_x = A^{(x+B)^C} + De^{(-E(log(x/F))^2)} + GH^x/(1+KGH^x)}

 \eqn{qx = 1-e^{(-HP_x)}}

This approximation ensures that qx, which is a probability, is in the correct range.

The lognormal model assumes that the log odds of death qx/(1-qx) has normal distribution
with a constant variance for all the ages. This model was proposed by Dellaportas et al.
(2001) and use the 8 parameters HP curve as follows:

 \eqn{log(qx/(1-qx)) = log(A^{(x+B)^C} + De^{-E(log(x/F))^2} + GH^x) + \epsilon_x}

Where \eqn{\epsilon_x} has independent distributions Normal(0, sigma2) for all ages. More details
of this model is avaible in Dellaportas, Smith e Stavropoulos (2001).

The reduced model does not consider the first term of the HP curve: \eqn{A^{(x+B)^C}}. In this
case, A, B and C are fixed as zero.

All parameters, with the exception of the extra parameter K of the Binomial and Poisson models
that is estimated by a optimization function, are estimated by the MCMC methods Gibbs sampling
for sigma2 and Metropolis-Hastings for parameters A, B, C, D, E, F, G and H. Informative prior
distributions should help the method to converge quickly.
}
\examples{
## Importing mortality data from USA
data(USA)

## Selecting the exposure and death count of the 2010 male population ranging from 0 to 90 years old
USA2010 = USA[USA$Year == 2010,]
x = 0:90
Ex = USA2010$Ex.Male[x+1]
Dx = USA2010$Dx.Male[x+1]

## Fitting binomial model
fit = hp(x = x, Ex = Ex, Dx = Dx)
print(fit)
fit$summary
fit$info

## Fitting lognormal model
## Specifying number of iterations, burn-in, thinning and the initial value of sigma2
fit = hp(x = x, Ex = Ex, Dx = Dx, model = "lognormal",
         M = 50000, bn = 10000, thin = 40, sigma2 = 0.1)
summary(fit)

## Fitting poisson model
## Specifying the prior distribution parameters for B and E and fixing K as 0.
fit = hp(x = x, Ex = Ex, Dx = Dx, model = "poisson",
         m = c(NA, 0.08, NA, NA, 9, NA, NA, NA),
         v = c(NA, 1e-4, NA, NA, 0.1, NA, NA, NA), K = 0)
summary(fit)

## Using other functions available in the package:
## plotting (See "?plot" in the BayesMortality package for more options):
plot(fit)

## qx estimation (See "?fitted" in the BayesMortality package for more options):
fitted(fit)

## chain's plot (See "?plot_chain" for more options):
plot_chain(fit)

## credible intervals (See "?qx_ci" for more options):
qx_ci(fit)


}
\references{
Dellaportas, P., Smith, A. F., and Stavropoulos, P. (2001). “Bayesian Analysis of Mortality Data.” \emph{Journal of the Royal Statistical Society: Series A (Statistics in Society)} 164 (2). Wiley Online Library: 275–291.

Heligman, Larry, and John H Pollard. (1980). “The Age Pattern of Mortality.” \emph{Journal of the Institute of Actuaries (1886-1994)} 107 (1). JSTOR: 49–80.
}
