% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_close.R
\name{hp_close}
\alias{hp_close}
\title{HP Model: Fitting the advanced ages of the life tables.}
\usage{
hp_close (fit, method = c("hp", "plateau", "linear", "gompertz"),
 x0 = max(fit$data$x), max_age = 120, k = 7,
 weights = seq(from = 0, to = 1, length.out = 2*k+1),
 new_Ex = NULL, new_Dx = NULL)
}
\arguments{
\item{fit}{Object of the class 'HP' adjusted by the hp() function}

\item{method}{Character string specifying the closing method to be fitted, with them being: 'hp', 'plateau', 'linear' or 'gompertz'.}

\item{x0}{Integer with the starting age the closing method will be fitted from. Default: Last age fitted by the 'HP' object.}

\item{max_age}{Integer with the maximum age the closing method will be fitted till. Default: 120 years old.}

\item{k}{Integer representing the size of the age-interval to be mixed with the 'linear' or 'gompertz' closing methods for a smooth graduation. If k = 0, no mixing will be made. Default: 7.}

\item{weights}{Vector of weights to be applied in the mixing of the life tables. Vector's size should be equal to 2*k+1.}

\item{new_Ex}{Exposure in the period after the x0 input. This argument is necessary for HP objects following the binomial and poisson distributions, as well as the 'linear' and 'gompertz' closing methods (optional).}

\item{new_Dx}{Vector containing the death counts in the period after the x0 input. This argument is necessary in the 'linear' and 'gompertz' closing methods, being equal in length with the new_Ex argument.}
}
\value{
Returns a 'ClosedHP' class object with the predictive chains of the death probability
(qx) from first fitted age to max_age argument, the data utilized by the function and the
closing method chosen.
}
\description{
This function receives an object of the class 'HP' adjusted by the hp() function
and fits a closing method to expand the life table's data to a maximum age argument inputed
by the user.
There are four closing methods to choose from: 'hp', 'plateau', 'linear' and 'gompertz'.
The 'linear' method can only be used with HP objects following the lognormal variant of
the HP mortality law.
}
\details{
The four closing methods implemented by the function are:
1.'hp' method: Expands the previously adjusted HP model till the max_age argument.
2.'plateau' method: Keeps the death probability (qx) constant after the x0 argument.
3.'linear' method: Fits a linear regression starting at age x0 - k till the last age with data available (lognormal only).
4.'gompertz' method: Used as the closing method of the 2010-2012 English Life Table No. 17, fits the Gompertz mortality law via SIR using the same available data as the 'linear' method.
}
\examples{
## Getting the mortality data of the United States provided by the Human Mortality Database (HMD):
data(USA)

## Selecting the exposure and the death count of the year 2010, ranging from 0 to 90 years old:
USA2010 = USA[USA$Year == 2010,]
x = 0:90
Ex = USA2010$Ex.Male[x+1]
Dx = USA2010$Dx.Male[x+1]

## Fitting a poisson and a lognormal model:
fit = hp(x = x, Ex = Ex, Dx = Dx,
         M = 40000, bn = 10000, thin = 30)
fit2 = hp(x = x, Ex = Ex, Dx = Dx, model = "lognormal",
          M = 40000, bn = 10000, thin = 30)

## Applying the closing function with different methods:
close1 = hp_close(fit2, method = "hp", x0 = 90)
close2 = hp_close(fit, method = "plateau", x0 = 90)
close3 = hp_close(fit2, method = "linear", x0 = 80,
                  new_Ex = USA2010$Ex.Male[82:101],
                  new_Dx = USA2010$Dx.Male[82:101])
close4 = hp_close(fit, method = "gompertz", x0 = 70,
                  new_Ex = USA2010$Ex.Male[72:101],
                  new_Dx = USA2010$Dx.Male[72:101],
                  k = 5, max_age = 120)

## Using the other functions available in the package with the 'ClosedHP' object:
qx_ci(close1)
fitted(close2)
expectancy(close3, age = 0:110)
plot(close4)
g <- plot(list(close4, fit),
          colors = c("seagreen", "blue"),
          labels = c("Fechamento", "Modelo"))
# plotly::ggplotly(g)

}
\references{
Dodd, Erengul, Forster, Jonathan, Bijak, Jakub, & Smith, Peter 2018. “Smoothing mortality data: the English life table, 2010-12.” \emph{Journal of the Royal Statistical Society: Series A (Statistics in Society)}, 181(3), 717-735.
}
