% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hp.R
\name{plot}
\alias{plot}
\alias{plot.HP}
\title{HP Model: Plotting the life tables.}
\usage{
plot(fits, age = NULL, Ex = NULL, plotIC = TRUE,
 plotData = TRUE, labels = NULL, colors = NULL,
 prob = 0.95)
}
\arguments{
\item{fits}{Object or a list of objects of the class 'HP' or 'ClosedHP' returned by the hp() or hp_close() functions.}

\item{age}{Vector with the ages to plot the life table(s).}

\item{Ex}{Vector with the exposures in the age period, having it's length equal to the age vector. This argument is only necessary when plotting poisson and/or binomial models.}

\item{plotIC}{Logical. If TRUE, plots the credible intervals (prediction intervals). Default: TRUE.}

\item{plotData}{Logical. If TRUE, plots the data used in the modelling as dots. Default: TRUE.}

\item{labels}{Vector with the name of the labels to be used in the legend, having it's length equal to the fits argument (Optional).}

\item{colors}{Vector with the colors to be used in the plot, having it's length equal to the fits argument (Optional).}

\item{prob}{Probability of the credible intervals. Default: 0.95}
}
\value{
A ggplot object with fitted life table(s).
}
\description{
Function that returns a log-scale ggplot of the 'HP' and 'ClosedHP' objects returned by the hp() and hp_close() functions.
}
\examples{
## Selecting the exposure and the death count of the year 1990, ranging from 0 to 90 years old:
USA1990 = USA[USA$Year == 1990,]
x = 0:90
Ex = USA1990$Ex.Male[x+1]
Dx = USA1990$Dx.Male[x+1]

## Fitting a poisson and a lognormal model:
fit = hp(x = x, Ex = Ex, Dx = Dx, model = "poisson",
         M = 40000, bn = 10000, thin = 30)
fit2 = hp(x = x, Ex = Ex, Dx = Dx, model = "lognormal",
          M = 40000, bn = 10000, thin = 30)

## Plotting the life tables:
plot(fit)
plot(fit2, age = 0:110, plotIC = TRUE)
plot(list(fit, fit2),
     age = 0:110, Ex = USA1990$Ex.Male,
     plotIC = FALSE, colors = c("red", "blue"),
     labels = c("Poisson", "Lognormal"))

}
