% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chain.R
\name{plot_chain}
\alias{plot_chain}
\title{HP Model: Chain's plot}
\usage{
plot_chain(fit, type = c("trace", "acf", "density"))
}
\arguments{
\item{fit}{Object of the class 'HP' fitted by the hp() function.}

\item{type}{Character string specifying the type of plot to be returned. There are three options:
\itemize{trace: }{return a plot with the parameters' samples.}
\itemize{acf: }{return a plot with the autocorrelation of the parameters' samples.}
\itemize{density: }{return a plot with the posterior density of the paramaters based on the samples generated by the MCMC method.}}
}
\value{
A plot of the chosen type.
}
\description{
This function provides three types of plots for the chain generated by the MCMC
method in the hp() function.
}
\examples{
## Importing mortality data from USA
data(USA)

## Selecting the exposure and death count of the 2010 population ranging from 0 to 90 years old
USA2010 = USA[USA$Year == 2010,]
x = 0:90
Ex = USA2010$Ex.Total[x+1]
Dx = USA2010$Dx.Total[x+1]

## Fitting a model
fit = hp(x = x, Ex = Ex, Dx = Dx, model = "lognormal",
         M = 50000, bn = 10000, thin = 40,
         m = c(NA, 0.08, rep(NA, 6)),
         v = c(NA, 1e-4, rep(NA, 6)))

## Plotting all the available types of plot
plot_chain(fit, type = "trace")
plot_chain(fit, type = "acf")
plot_chain(fit, type = "density")

}
