% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_ci.R
\name{qx_ci}
\alias{qx_ci}
\title{HP Model: Credible intervals.}
\usage{
qx_ci(fit, age = NULL, Ex = NULL, prob = 0.95)
}
\arguments{
\item{fit}{Object of the class 'HP' or 'ClosedHP' adjusted by the hp() or hp_close() functions.}

\item{age}{Vector with the ages to calculate the credible intervals (Optional).}

\item{Ex}{Vector with the exposures in the age period, having it's length equal to the age vector. This argument is only necessary when using poisson and binomial models.}

\item{prob}{Probability of the credible intervals. Default: 0.95}
}
\value{
A data.frame object with the selected ages and the corresponding credible intervals for the death probabilities.
}
\description{
This function returns the credible intervals (prediction intervals) of the death probabilities of a 'HP' or 'ClosedHP' class object adjusted by the hp() or hp_close() functions.
}
\examples{
## Getting the mortality data of the United States provided by the Human Mortality Database (HMD):
data(USA)

## Selecting the exposure and the death count of the year 2000, ranging from 0 to 90 years old:
USA2000 = USA[USA$Year == 2000,]
x = 0:90
Ex = USA2000$Ex.Total[x+1]
Dx = USA2000$Dx.Total[x+1]

## Fitting a poisson and a lognormal model:
fit = hp(x = x, Ex = Ex, Dx = Dx, model = "poisson",
         M = 40000, bn = 10000, thin = 30)
fit2 = hp(x = x, Ex = Ex, Dx = Dx, model = "lognormal",
          M = 40000, bn = 10000, thin = 30)

## credible intervals (prediction intervals)
qx_ci(fit)
qx_ci(fit, age = 0:110, Ex = USA2000$Ex.Total)
qx_ci(fit2, age = 0:110, prob = 0.9)

}
